requirements.txt
asgiref==3.2.3
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dockerpty==0.4.1
idna==2.8
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
PyYAML==3.13
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
texttable==0.9.1
urllib3==1.25.7
virtualenv==16.7.9
